refactor: extract useLocalStorage hook
